refactor: extract interface from `HttpFactory`
