+++ src/Bootstrap.php
@@ -11,6 +11,8 @@ use Internal\DLoad\Module\Common\Internal\Injection\ConfigLoader;
 use Internal\DLoad\Module\Common\Internal\ObjectContainer;
 use Internal\DLoad\Module\Common\OperatingSystem;
 use Internal\DLoad\Module\Common\Stability;
+use Internal\DLoad\Module\HttpClient\Factory;
+use Internal\DLoad\Module\HttpClient\Internal\NyholmFactoryImpl;
 use Internal\DLoad\Module\Repository\Internal\GitHub\Factory as GithubRepositoryFactory;
 use Internal\DLoad\Module\Repository\RepositoryProvider;
 use Internal\DLoad\Service\Container;
@@ -105,6 +107,10 @@ final class Bootstrap
         $this->container->bind(
             BinaryProvider::class,
             static fn(Container $c): BinaryProvider => $c->get(BinaryProviderImpl::class),
+        );
+        $this->container->bind(
+            Factory::class,
+            static fn(Container $c): Factory => $c->get(NyholmFactoryImpl::class),
         );
 
         return $this;
